server/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
pydantic-settings==2.8.1
